ECO-822: fix circularly importing issue
